[file/#12]:경로,파일명 수정
