--- src/app/__home/page.jsx
@@ -1,10 +1,0 @@
-import Header from "@/src/ui/layout/Header";
-import BlueBox from "@/src/ui/Components/home/BlueBox";
-
-export default function MainHome() {
-  return (
-    <div className={`flex flex-col justify-center items-center h-full `}>
-      <BlueBox />
-    </div>
-  );
-}
